+++ pkg/organizations/README.md
@@ -1,0 +1,15 @@
+# organizations
+
+[![Go Report Card](https://img.shields.io/badge/dev-reference-007d9c?logo=go&logoColor=white&style=flat)](https://pkg.go.dev/github.com/workos-inc/workos-go/pkg/organizations)
+
+A Go package to make requests to the WorkOS Organizations API.
+
+## Install
+
+```sh
+go get -u github.com/workos-inc/workos-go/pkg/organizations
+```
+
+## How it works
+
+See the [Organizations API reference](https://workos.com/docs/reference/organization).
